% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{apply_data_set_up}
\alias{apply_data_set_up}
\title{Applies grouping to data set conditionally}
\usage{
apply_data_set_up(df, group)
}
\arguments{
\item{df}{data frame}

\item{group}{supply reactive output from group selector}
}
\value{
returns df either grouped or not
}
\description{
Applies grouping to data set conditionally
}
