% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_limits_per_groups}
\alias{calc_limits_per_groups}
\title{Return x and y limits of "group-subsetted" dframe}
\usage{
calc_limits_per_groups(dframe, group_index, xvar, yvar, scaling = 0.02)
}
\arguments{
\item{dframe}{dataframe/tibble, grouped/ungrouped}

\item{group_index}{numeric, group indices for which to return lims}

\item{xvar}{character, name of x var for plot (must exist in dframe)}

\item{yvar}{character, name of y var for plot (must exist in dframe)}

\item{scaling}{numeric, 1 +/- \code{scaling} times limits}
}
\value{
list with xlim and ylim
}
\description{
Used for adjusting layout of plotly plot based on selected
groups in \code{group_selector_table}; currently used in viz tab
}
