% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{can_internet}
\alias{can_internet}
\title{Check for internet connection}
\usage{
can_internet(url = "http://www.google.com")
}
\arguments{
\item{url}{character, valid path to url - user responsible}
}
\value{
logical - TRUE or FALSE
}
\description{
Check for internet connection
}
