% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_active_filters}
\alias{check_active_filters}
\title{Identify filter inputs by id and label}
\usage{
check_active_filters(allinputs)
}
\arguments{
\item{allinputs}{list of all currently active inputs}
}
\value{
list with id of last filter (character), and number of last filter (numeric)
}
\description{
Identify filter inputs by id and label
}
