% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_individual_statement}
\alias{check_individual_statement}
\title{check if a filter statement is valid}
\usage{
check_individual_statement(df, statement)
}
\arguments{
\item{df}{data frame / tibble to be filtered}

\item{statement}{character string,}
}
\value{
logical, did filter statement work?
}
\description{
check if a filter statement is valid
}
