% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checked_filter}
\alias{checked_filter}
\title{Filter a data frame with series of statements}
\usage{
checked_filter(df, statements, apply_grouped)
}
\arguments{
\item{df}{data frame / tibble to be filtered}

\item{statements}{character vector of individual conditional statements; need not evaluate successfully individually}

\item{apply_grouped}{logical, should filter be applied to each group?}
}
\value{
list, logical vector of success and failures, and
}
\description{
Filter a data frame with series of statements
}
