% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_missing_data.R
\name{clean_missing_data}
\alias{clean_missing_data}
\title{Clean Missing Data}
\usage{
clean_missing_data(data, method = "remove")
}
\arguments{
\item{data}{A data frame.}

\item{method}{A character string specifying the method to handle missing values. Options are "remove" or "mean".}
}
\value{
A data frame with missing values handled.
}
\description{
This function handles missing values in a dataset by either removing them or filling them using a specified method.
}
\examples{
data <- data.frame(a = c(1, 2, NA, 4), b = c(NA, 2, 3, 4))
clean_missing_data(data, method = "mean")
}
