% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{col2plotlyrgb}
\alias{col2plotlyrgb}
\title{Color conversion for plotly with alpha}
\usage{
col2plotlyrgb(colorname)
}
\arguments{
\item{colorname}{string, hex value or R color name}
}
\value{
string, length 1, with rgb code for plotly
}
\description{
Color conversion for plotly with alpha
}
