% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{col2plotlyrgba}
\alias{col2plotlyrgba}
\title{Color conversion for plotly with alpha}
\usage{
col2plotlyrgba(colorname, alpha)
}
\arguments{
\item{colorname}{string, hex value or R color name}

\item{alpha}{numeric, value for alpha (transparency)}
}
\value{
string, length 1, with rgb code for plotly
}
\description{
Color conversion for plotly with alpha
}
