% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive-stats-utils.R
\name{compute_normality}
\alias{compute_normality}
\title{compute the normality of a data column with the shapiro test}
\usage{
compute_normality(data, var, data_type, output)
}
\arguments{
\item{data}{Tibble from spreadsheet}

\item{var}{The column name in data to look at}

\item{data_type}{String, "continuous", "ordinal", "categorical", "binary"}

\item{output}{something to do with the shapiro test}
}
\value{
return something
}
\description{
compute the normality of a data column with the shapiro test
}
