% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_iso8601.R
\docType{data}
\name{data_iso8601}
\alias{data_iso8601}
\title{Example data for test driving iso8601_* functions}
\format{
A data frame with 31 rows and 4 columns:
\describe{
  \item{date}{Time series incrementing by day}
  \item{date_2_formats}{Time series incrementing by day in two formats}
  \item{datetime}{Dates and times incrementing by hour}
  \item{datetime_mess}{Single date and time represented in many formats}
}
}
\usage{
data_iso8601
}
\description{
A dataset containing dates and times of different formats.
}
\keyword{datasets}
