% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacleanr-package.R
\docType{package}
\name{datacleanr-package}
\alias{datacleanr}
\alias{datacleanr-package}
\title{datacleanr: Interactive and Reproducible Data Cleaning}
\description{
Flexible and efficient cleaning of data with interactivity. 'datacleanr' facilitates best practices in data analyses and reproducibility with built-in features and by translating interactive/manual operations to code. The package is designed for interoperability, and so seamlessly fits into reproducible analyses pipelines in 'R'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/the-Hull/datacleanr}
  \item Report bugs at \url{https://github.com/the-Hull/datacleanr/issues}
}

}
\author{
\strong{Maintainer}: Alexander Hurley \email{agl.hurley@gmail.com} (\href{https://orcid.org/0000-0002-9641-2805}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Richard Peters (\href{https://orcid.org/0000-0002-7441-1297}{ORCID}) [contributor]
  \item Christoforos Pappas (\href{https://orcid.org/0000-0001-5721-557X}{ORCID}) [contributor]
}

}
\keyword{internal}
