% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacleanr_main.R
\name{datacleanr_module}
\alias{datacleanr_module}
\title{Data cleanr}
\usage{
datacleanr_module(dframe, viewer = "browser")
}
\arguments{
\item{dframe}{dataframe}

\item{viewer}{character, "browser", "pane" for respective viewers; any other string will open the RStudio dialaogue view}
}
\description{
Data cleanr
}
