% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacleanr_app.R
\name{dcr_app}
\alias{dcr_app}
\title{Interactive data cleaning}
\usage{
dcr_app(dframe, browser = FALSE)
}
\arguments{
\item{dframe}{dataframe, tbl or data.table}

\item{browser}{logical, should app start in OS's default browser?}
}
\description{
Interactive data cleaning
}
\details{
\code{datacleanr} provides an interactive data overview, and allows
reproducible subsetting, and reproducible visual outlier detection and annotation.
The app works with \code{data.frame}, \code{tibble} and \code{data.table} objects.
Note, that spatial data is supported; this requires that \code{lon} and \code{lat} columns in degrees.
are present.
The interactive visualization relies on subsetting (i.e. highlighting) sub-groups of the provided data.
These groups are composed of unique combinations of columns in the data set (must be \code{factor}),
and are carried through the app for exploratory analyses (tab \strong{Overview}), and plotting
(tab \strong{Visualization}).
The tab \strong{Extraction} provides code to reproduce the entire procedure, which can be copied,
or sent directly to an active \code{RStudio} script.
}
