% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dcr_checks}
\alias{dcr_checks}
\title{Initial checks for data set}
\usage{
dcr_checks(dframe)
}
\arguments{
\item{dframe}{dframe supplied to \code{dcr_app}}
}
\value{
Nothing - just check!
}
\description{
Initial checks for data set
}
