% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extend_palette}
\alias{extend_palette}
\title{Identify filter inputs by id and label}
\usage{
extend_palette(n)
}
\arguments{
\item{n}{numeric, number of colors}

\item{allinputs}{list of all currently active inputs}

\item{df}{dataframe to be filtered}

\item{statements}{character, vector with individual statements for filtering}
}
\value{
list with id of last filter (character), and number of last filter (numeric)

check if filtering statement is successfull

logical vector of fails and successes

extend brewer palette

color vector of length n
}
\description{
Identify filter inputs by id and label
}
