% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_scoped}
\alias{filter_scoped}
\title{Apply filter based on a statement, scoped to \code{dplyr} groups}
\usage{
filter_scoped(dframe, statement, scope_at)
}
\arguments{
\item{dframe}{data.frame/tbl, grouped or ungrouped}

\item{statement}{character, statement for filtering (only VALID expressions; use \code{check_individual_statement} to grab only valid.}

\item{scope_at}{numeric, group indices to apply filter statements to}
}
\value{
List, containing item \code{filtered_df}, a \code{data.frame} filtered based on statements and scope.
}
\description{
Apply filter based on a statement, scoped to \code{dplyr} groups
}
