% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orchestration.R
\name{finalize_salary_tbl}
\alias{finalize_salary_tbl}
\title{Create a salary table object}
\usage{
finalize_salary_tbl(data)
}
\arguments{
\item{data}{a data frame containing salary data}
}
\value{
An object of class "salary_tbl"
}
\description{
Create a salary table object
}
