% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_replace.R
\name{find_replace.default}
\alias{find_replace.default}
\title{Renames the values on an input according to a pattern-value map}
\usage{
\method{find_replace}{default}(x, ...)
}
\arguments{
\item{x}{Argument of a type for which no find_replace method exists yet.}

\item{...}{arguments passed on to other methods}
}
\description{
Fallback method
}
