% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_replace.R
\name{find_replace}
\alias{find_replace}
\title{Renames the values on an input according to a pattern-value map}
\usage{
find_replace(...)
}
\arguments{
\item{...}{arguments passed to other methods}
}
\description{
Renames the values on an input according to a pattern-value map
}
\seealso{
\code{\link{find_replace.data.frame}}
}
