% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_factor_cols_idx}
\alias{get_factor_cols_idx}
\title{Identify columns carrying non-numeric values}
\usage{
get_factor_cols_idx(x)
}
\arguments{
\item{x}{data.frame}
}
\value{
logical, is column in x non-numeric?
}
\description{
Identify columns carrying non-numeric values
}
