% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{handle_add_outlier_trace}
\alias{handle_add_outlier_trace}
\title{Handle Add traces}
\usage{
handle_add_outlier_trace(
  sp,
  dframe,
  ok,
  selectors,
  trace_map,
  source = "scatterselect",
  session
)
}
\arguments{
\item{sp}{selected points}

\item{dframe}{plot data}

\item{ok}{reactive, old keys}

\item{selectors}{reactive input selectors}

\item{trace_map}{numeric, previous max trace id}

\item{source}{plotly source}

\item{session}{active session}
}
\description{
Handle Add traces
}
