% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{handle_add_traces}
\alias{handle_add_traces}
\title{Handle Add traces}
\usage{
handle_add_traces(sp, dframe, ok, selectors, source = "scatterselect", session)
}
\arguments{
\item{sp}{selected points}

\item{dframe}{plot data}

\item{ok}{reactive, old keys}

\item{selectors}{reactive input selectors}

\item{source}{plotly source}

\item{session}{active session}
}
\description{
Handle Add traces
}
