% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{handle_outlier_selection}
\alias{handle_outlier_selection}
\title{Handle selections/clicks of outliers}
\usage{
handle_outlier_selection(sel_data_old, sel_data_new)
}
\arguments{
\item{sel_data_old}{\code{data.frame}, which must have columns \code{keys}, \code{selection_count} and \code{.annotation}}

\item{sel_data_new}{\code{data.frame}, from \code{plotly::event_data}, with column \code{customData} as \code{.dcrkey}}
}
\value{
\code{data.frame}, to update the input of \code{sel_data_old}.
}
\description{
Handle selections/clicks of outliers
}
