% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{handle_sel_outliers}
\alias{handle_sel_outliers}
\title{Handle selection of outliers (with select - unselect capacity)}
\usage{
handle_sel_outliers(sel_old_df, sel_new)
}
\arguments{
\item{sel_old_df}{data.frame of selection info}

\item{sel_new}{data.frame, event data from plotly, must have column \code{customdata}}
}
\value{
updated selection data frame
}
\description{
Handle selection of outliers (with select - unselect capacity)
}
