% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{handle_selection}
\alias{handle_selection}
\title{handle plotly selections}
\usage{
handle_selection(old, new)
}
\arguments{
\item{old}{numeric or NULL}

\item{new}{numeric or NULL}
}
\value{
numeric or NULL
}
\description{
handle plotly selections
}
