% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_module.R
\name{handle_selections}
\alias{handle_selections}
\title{Handle point selection svr side}
\usage{
handle_selections(input, output, session)
}
\arguments{
\item{input}{Shiny}

\item{output}{Shiny}

\item{session}{Shiny}
}
\value{
selected points
}
\description{
Handle point selection svr side
}
