% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hide_trace_idx}
\alias{hide_trace_idx}
\title{Provide traces ids to set to invisible}
\usage{
hide_trace_idx(max_groups, selected_groups)
}
\arguments{
\item{max_groups}{numeric, number of groups in grouptable}

\item{selected_groups}{groups highlighted in grouptable}
}
\description{
Provide traces ids to set to invisible
}
\details{
Provides the indices (JS notation, starting at 0) for indices
that are set to \code{visible = 'legendonly'} through \code{plotly.restyle}
}
