% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_na.R
\name{impute_na}
\alias{impute_na}
\title{Impute missing values}
\usage{
impute_na(x, method = "mean")
}
\arguments{
\item{x}{Numeric vector}

\item{method}{Method: "mean", "median" or numeric value}
}
\value{
Numeric vector with imputed values
}
\description{
Impute missing values
}
\examples{
impute_na(c(1,NA,3), method="mean")
impute_na(c(1,NA,3), method=0)
}
