% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso8601_format.R
\name{iso8601_format}
\alias{iso8601_format}
\title{Get datetime format string from ISO 8601 datetimes}
\usage{
iso8601_format(x)
}
\arguments{
\item{x}{(character) A vector of dates and times created with
`EDIutils::iso8601_char`.}
}
\value{
(character) A datetime format string if only one is found, otherwise
    a vector of datetime format strings are returned.
}
\description{
Get format of ISO 8601 formatted character string.
}
