% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_group_table}
\alias{make_group_table}
\title{Make grouping overview table}
\usage{
make_group_table(dframe)
}
\arguments{
\item{dframe}{data.frame}
}
\value{
tibble with one row per group
}
\description{
Make grouping overview table
}
