% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_save_filepath}
\alias{make_save_filepath}
\title{Wrapper for saving files}
\usage{
make_save_filepath(save_dir, input_filepath, suffix, ext)
}
\arguments{
\item{save_dir}{character, selected save dir}

\item{input_filepath}{character, original file path to folder}

\item{suffix}{character, e.g. 'CLEAN' or 'cleaning_script'}

\item{ext}{character, file extension, no dot!!}
}
\value{
OS-conform file path for saving
}
\description{
Wrapper for saving files
}
