% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mapping.R
\docType{data}
\name{size_mapping}
\alias{size_mapping}
\alias{size_levels}
\title{Company size mappings}
\format{
Named character vector (\code{size_mapping}) and character vector of size
labels (\code{size_levels}).

An object of class \code{character} of length 3.
}
\usage{
size_mapping

size_levels
}
\description{
Normalised size buckets for companies.
}
\keyword{datasets}
