% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mapping.R
\docType{data}
\name{currencies_levels}
\alias{currencies_levels}
\alias{exchange_rates_to_usd}
\title{Currency mappings and FX table}
\format{
\itemize{
\item \code{currencies_levels}: vector of supported currency codes.
\item \code{exchange_rates_to_usd}: data frame with \code{year}, \code{currency} and \code{rate}
columns used by \code{convert_currency_to_usd()}.
}

An object of class \code{data.frame} with 164 rows and 3 columns.
}
\usage{
currencies_levels

exchange_rates_to_usd
}
\description{
Reference currency codes and yearly exchange rates to USD.
}
\keyword{datasets}
