% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mapping.R
\docType{data}
\name{mapping_employement_type}
\alias{mapping_employement_type}
\alias{levels_employement_type}
\title{Employment type mappings}
\format{
Named character vector (\code{mapping_employement_type}) and character
vector of possible levels (\code{levels_employement_type}).

An object of class \code{character} of length 5.
}
\usage{
mapping_employement_type

levels_employement_type
}
\description{
Normalised employment types and their codes.
}
\keyword{datasets}
