% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mapping.R
\docType{data}
\name{experience_mapping}
\alias{experience_mapping}
\alias{experience_labels_ordered}
\title{Experience level mappings}
\format{
Named character vector (\code{experience_mapping}) and ordered levels
(\code{experience_labels_ordered}).

An object of class \code{character} of length 4.
}
\usage{
experience_mapping

experience_labels_ordered
}
\description{
Normalised experience levels used in salary datasets.
}
\keyword{datasets}
