% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mapping.R
\docType{data}
\name{levels_job_title}
\alias{levels_job_title}
\alias{mapping_job_title}
\title{Job title mappings}
\format{
Character vectors:
\itemize{
\item \code{levels_job_title}: the target categories.
\item \code{mapping_job_title}: a named vector mapping raw titles to
\code{levels_job_title}.
}

An object of class \code{character} of length 93.
}
\usage{
levels_job_title

mapping_job_title
}
\description{
Reference vectors used to normalise raw job titles into a limited set of
harmonised categories.
}
\keyword{datasets}
