% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mapping.R
\docType{data}
\name{region_map}
\alias{region_map}
\alias{regions_levels}
\alias{levels_iso2}
\alias{mapping_total}
\title{Location mappings}
\format{
Character vectors and named vectors:
\itemize{
\item \code{region_map}: map ISO-2 codes to broad regions.
\item \code{regions_levels}: ordered list of region labels.
\item \code{levels_iso2}: ISO-2 codes expected by the package.
\item \code{mapping_total}: mapping of ISO-2 codes to themselves (used for
\code{normalize_factor()}).
}

An object of class \code{character} of length 6.

An object of class \code{character} of length 106.

An object of class \code{character} of length 222.
}
\usage{
region_map

regions_levels

levels_iso2

mapping_total
}
\description{
ISO-2 country mappings, regions and normalised lookup tables used for
location cleaning.
}
\keyword{datasets}
