% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_apply_reset.R
\name{module_server_apply_reset}
\alias{module_server_apply_reset}
\title{Server Module: apply / reset filter}
\usage{
module_server_apply_reset(input, output, session, df_filtered, df_original)
}
\arguments{
\item{input, output, session}{standard}
}
\description{
Server Module: apply / reset filter
}
