% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_str_select_box.R
\name{module_server_box_str_filter}
\alias{module_server_box_str_filter}
\title{Server Module: box for str filter condition}
\usage{
module_server_box_str_filter(
  input,
  output,
  session,
  selector = "h2",
  actionbtn
)
}
\arguments{
\item{input, output, session}{standard}
}
\description{
Server Module: box for str filter condition
}
