% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_render_checkbox.R
\name{module_server_checkbox}
\alias{module_server_checkbox}
\title{Server Module: checkbox rendering}
\usage{
module_server_checkbox(input, output, session, text)
}
\arguments{
\item{input, output, session}{standard \code{shiny} boilerplate}

\item{text}{Character, appears next to checkbox (or coerced)}
}
\description{
Server Module: checkbox rendering
}
