% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_filter_df.R
\name{module_server_df_filter}
\alias{module_server_df_filter}
\title{Server Module: filter info text and filtered df output}
\usage{
module_server_df_filter(input, output, session, df, statements)
}
\arguments{
\item{input, output, session}{standard \code{shiny} boilerplate}

\item{df}{data frame/tibble for filtering}

\item{statements}{character, lengths >= 0, with (valid) statements for \code{\link{base::subset}}}
}
\value{
df, either filtered or original, based on validity of \code{statements}
}
\description{
Server Module: filter info text and filtered df output
}
