% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_extract_code_box.R
\name{module_server_extract_code}
\alias{module_server_extract_code}
\title{Server Module:  Selection Annotator}
\usage{
module_server_extract_code(
  input,
  output,
  session,
  df_label,
  filter_df,
  gvar,
  statements,
  sel_points,
  overwrite
)
}
\arguments{
\item{input, output, session}{standard \code{shiny} boilerplate}

\item{df_label}{string, name of original df input}

\item{filter_df}{data frame with filter statements and scoping lvl}

\item{gvar}{character, grouping vars for \code{dplyr::group_by}}

\item{statements}{lgl, vector of working statements}

\item{sel_points}{data frame with selected point keys, annotations, and selection count}

\item{overwrite}{reacive value, TRUE/FALSE from checkbox input}
}
\description{
Server Module:  Selection Annotator
}
