% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_extract_code_fileconfig.R
\name{module_server_extract_code_fileconfig}
\alias{module_server_extract_code_fileconfig}
\title{Server Module: Extraction File selection menu}
\usage{
module_server_extract_code_fileconfig(
  input,
  output,
  session,
  df_label,
  is_on_disk,
  has_processed
)
}
\arguments{
\item{input, output, session}{standard \code{shiny} boilerplate}

\item{df_label}{character, name of original df input}

\item{is_on_disk}{Logical, whether df represented by \code{df_label} was on disk or from interactive \code{R} use}

\item{has_processed}{reactive, logical, TRUE if filtered / selected points}
}
\description{
Server Module: Extraction File selection menu
}
