% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_grouptable_relayout.R
\name{module_server_group_relayout_buttons}
\alias{module_server_group_relayout_buttons}
\title{Server Module:  Selection Annotator}
\usage{
module_server_group_relayout_buttons(input, output, session, startscatter)
}
\arguments{
\item{input, output, session}{standard \code{shiny} boilerplate}

\item{startscatter}{reactive, actionbutton value}
}
\value{
reactive values with input xvar, yvar and actionbutton counter
}
\description{
Server Module:  Selection Annotator
}
\details{
provides UI text box element
}
