% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_set_grouping_selection.R
\name{module_server_group_select}
\alias{module_server_group_select}
\title{Server Module: group selection}
\usage{
module_server_group_select(input, output, session, dframe)
}
\arguments{
\item{input, output, session}{standard}

\item{dframe}{data frame for filtering}
}
\description{
Server Module: group selection
}
