% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_dynamic_histograms.R
\name{module_server_histograms}
\alias{module_server_histograms}
\title{Server Module: dynamic histogram output for n vars
str filter condition}
\usage{
module_server_histograms(
  input,
  output,
  session,
  dframe,
  dframe_recover,
  selector_inputs,
  sel_points
)
}
\arguments{
\item{input, output, session}{standard \code{shiny} boilerplate}

\item{dframe}{df}

\item{dframe_recover}{df, unfiltered df (faster subsetting via dcrkeys)}

\item{selector_inputs}{reactive vals from above-plot controls,}

\item{sel_points}{reactive, provides .dcrkey of selected points}
}
\value{
reactive values with input xvar, yvar and actionbutton counter
}
\description{
Server Module: dynamic histogram output for n vars
str filter condition
}
\details{
provides UI buttons for deleting last / entire outlier selection
}
