% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_plot_lowercontrol_buttons.R
\name{module_server_lowercontrol_btn}
\alias{module_server_lowercontrol_btn}
\title{Server Module: box for str filter condition}
\usage{
module_server_lowercontrol_btn(
  input,
  output,
  session,
  selector_inputs,
  action_track
)
}
\arguments{
\item{input, output, session}{standard \code{shiny} boilerplate}

\item{selector_inputs}{reactive vals from above-plot controls, used to determine if plot is a map (lon/lat)}

\item{action_track}{reactive, logical - has plot been pressed?}
}
\value{
reactive values with input xvar, yvar and actionbutton counter
}
\description{
Server Module: box for str filter condition
}
\details{
provides UI buttons for deleting last / entire outlier selection
}
