% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_annotation_table.R
\name{module_server_plot_annotation_table}
\alias{module_server_plot_annotation_table}
\title{Server Module: DT for annotation}
\usage{
module_server_plot_annotation_table(input, output, session, dframe, sel_points)
}
\arguments{
\item{input, output, session}{standard \code{shiny} boilerplate}

\item{dframe}{df used for plotting}

\item{sel_points}{numeric, vector of .dcrkeys selected in plot}
}
\value{
df with .dcrkeys and annotations
}
\description{
Server Module: DT for annotation
}
