% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_plot_selectable.R
\name{module_server_plot_selectable}
\alias{module_server_plot_selectable}
\title{Server Module: box for str filter condition}
\usage{
module_server_plot_selectable(
  input,
  output,
  session,
  df,
  group_row,
  selector_inputs,
  sel_points
)
}
\arguments{
\item{input, output, session}{standard \code{shiny} boilerplate}

\item{df}{reactive df, with df as element}

\item{group_row}{numeric, selected group_index}

\item{selector_inputs}{reactive, output from module_plot_selectorcontrols}

\item{sel_points}{numeric, provides .dcrkey of selected points}
}
\description{
Server Module: box for str filter condition
}
\details{
provides plot, note, that data set needs a column .dcrkey, added in initial processing step
}
