% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_plot_selectorcontrols.R
\name{module_server_plot_selectorcontrols}
\alias{module_server_plot_selectorcontrols}
\title{Server Module: box for str filter condition}
\usage{
module_server_plot_selectorcontrols(input, output, session, df)
}
\arguments{
\item{input, output, session}{standard \code{shiny} boilerplate}

\item{df}{reactive df, with df as element}
}
\value{
reactive values with input xvar, yvar and actionbutton counter
}
\description{
Server Module: box for str filter condition
}
\details{
provides UI text box element
}
