% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_annotation_button.R
\name{module_server_text_annotator}
\alias{module_server_text_annotator}
\title{Server Module:  Selection Annotator}
\usage{
module_server_text_annotator(input, output, session, sel_data)
}
\arguments{
\item{input, output, session}{standard \code{shiny} boilerplate}

\item{sel_data}{reactive df}
}
\value{
reactive values with input xvar, yvar and actionbutton counter
}
\description{
Server Module:  Selection Annotator
}
\details{
provides UI text box element
}
