% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_apply_reset.R
\name{module_ui_apply_reset}
\alias{module_ui_apply_reset}
\title{UI Module: Apply/Reset Filtering}
\usage{
module_ui_apply_reset(id)
}
\arguments{
\item{id}{Character, identifier for variable selection}
}
\description{
UI Module: Apply/Reset Filtering
}
