% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_render_checkbox.R
\name{module_ui_checkbox}
\alias{module_ui_checkbox}
\title{UI Module: data summary}
\usage{
module_ui_checkbox(id, cond_id)
}
\arguments{
\item{id}{shiny standard}

\item{cond_id}{character,}
}
\description{
UI Module: data summary
}
