% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_filter_df.R
\name{module_ui_df_filter}
\alias{module_ui_df_filter}
\title{UI Module: filter info text output}
\usage{
module_ui_df_filter(id)
}
\arguments{
\item{id}{character, shiny namespacing}
}
\value{
UI text element giving number of failed filters and percent of filtered rows
}
\description{
UI Module: filter info text output
}
