% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_extract_code_box.R
\name{module_ui_extract_code}
\alias{module_ui_extract_code}
\title{UI Module: Extraction Text output}
\usage{
module_ui_extract_code(id)
}
\arguments{
\item{id}{Character string}
}
\description{
UI Module: Extraction Text output
}
