% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_extract_code_fileconfig.R
\name{module_ui_extract_code_fileconfig}
\alias{module_ui_extract_code_fileconfig}
\title{UI Module: Extraction File selection menu}
\usage{
module_ui_extract_code_fileconfig(id)
}
\arguments{
\item{id}{Character string}
}
\description{
UI Module: Extraction File selection menu
}
