% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_str_select_box.R
\name{module_ui_filter_str}
\alias{module_ui_filter_str}
\title{UI Module: box for str filter condition}
\usage{
module_ui_filter_str(id)
}
\arguments{
\item{id}{Character string}
}
\description{
UI Module: box for str filter condition
}
