% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_grouptable_relayout.R
\name{module_ui_group_relayout_buttons}
\alias{module_ui_group_relayout_buttons}
\title{UI Module: Grouptable Relayout Buttons}
\usage{
module_ui_group_relayout_buttons(id)
}
\arguments{
\item{id}{Character string}
}
\description{
UI Module: Grouptable Relayout Buttons
}
