% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_group_selection.R
\name{module_ui_group_select}
\alias{module_ui_group_select}
\title{UI Module: group selection}
\usage{
module_ui_group_select(df, id)
}
\arguments{
\item{df}{data.frame loaded into gadget; should support df, tibble, data.table}

\item{id}{Character, identifier for variable selection}
}
\description{
UI Module: group selection
}
