% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_group_selector_table.R
\name{module_ui_group_selector_table}
\alias{module_ui_group_selector_table}
\title{UI Module: box for str filter condition}
\usage{
module_ui_group_selector_table(id)
}
\arguments{
\item{id}{Character string}
}
\description{
UI Module: box for str filter condition
}
