% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_dynamic_histograms.R
\name{module_ui_histograms}
\alias{module_ui_histograms}
\title{UI Module: dynamic histogram output for n vars}
\usage{
module_ui_histograms(id)
}
\arguments{
\item{id}{Character string}
}
\description{
UI Module: dynamic histogram output for n vars
}
