% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_plot_lowercontrol_buttons.R
\name{module_ui_lowercontrol_btn}
\alias{module_ui_lowercontrol_btn}
\title{UI Module: Delete selection buttons}
\usage{
module_ui_lowercontrol_btn(id)
}
\arguments{
\item{id}{Character string}
}
\description{
UI Module: Delete selection buttons
}
