% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_annotation_table.R
\name{module_ui_plot_annotation_table}
\alias{module_ui_plot_annotation_table}
\title{UI Module: DT for annotation}
\usage{
module_ui_plot_annotation_table(id)
}
\arguments{
\item{id}{Character string}
}
\description{
UI Module: DT for annotation
}
