% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_plot_selectable.R
\name{module_ui_plot_selectable}
\alias{module_ui_plot_selectable}
\title{UI Module: plotly plot}
\usage{
module_ui_plot_selectable(id)
}
\arguments{
\item{id}{Character string}
}
\description{
UI Module: plotly plot
}
