% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_plot_selectorcontrols.R
\name{module_ui_plot_selectorcontrols}
\alias{module_ui_plot_selectorcontrols}
\title{UI Module: selector controls}
\usage{
module_ui_plot_selectorcontrols(id)
}
\arguments{
\item{id}{Character string}
}
\description{
UI Module: selector controls
}
