% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_annotation_button.R
\name{module_ui_text_annotator}
\alias{module_ui_text_annotator}
\title{UI Module: Selection Annotator}
\usage{
module_ui_text_annotator(id)
}
\arguments{
\item{id}{Character string}
}
\description{
UI Module: Selection Annotator
}
