% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalisation.R
\name{normalize_experience_level}
\alias{normalize_experience_level}
\title{Normalize an ordered factor column}
\usage{
normalize_experience_level(data, column_name, levels_ordered)
}
\arguments{
\item{data}{data.frame A data frame containing the column to normalize}

\item{column_name}{character Name of the column to normalize}

\item{levels_ordered}{character Vector of levels in the desired order}
}
\value{
data.frame The same data frame with the specified column converted to an ordered factor
}
\description{
Normalize an ordered factor column
}
