% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.dcr_code}
\alias{print.dcr_code}
\title{Method for printing dcr_code output}
\usage{
\method{print}{dcr_code}(x, ...)
}
\arguments{
\item{x}{character, code  output from \code{dcr_app}}
}
\description{
Method for printing dcr_code output
}
