% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_zero_na.R
\name{replace_zero_na}
\alias{replace_zero_na}
\title{Replace zeros with NAs}
\usage{
replace_zero_na(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A numeric vector with zeros replaced by NAs
}
\description{
This function replaces all zeros in a numeric vector with NA values.
Useful for sensor data where zero means missing.
}
\examples{
replace_zero_na(c(0, 1, 2, 0, 3))

}
