% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_module.R
\name{scatter_plot_svr_module}
\alias{scatter_plot_svr_module}
\title{Make svr-side plot}
\usage{
scatter_plot_svr_module(input, output, session, data, vals, x, y)
}
\arguments{
\item{input}{Shiny basics}

\item{output}{Shiny basics}

\item{session}{Shiny basics}

\item{data}{data set passed to gadget app}

\item{vals}{reactive vals used for subsetting}
}
\value{
selected vals
}
\description{
Make svr-side plot
}
