% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_module.R
\name{scatter_plot_ui_module}
\alias{scatter_plot_ui_module}
\title{Make ui plot}
\usage{
scatter_plot_ui_module(id)
}
\arguments{
\item{id}{Character, label for UI}
}
\value{
UI plotoutput
}
\description{
Make ui plot
}
