% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_groups}
\alias{split_groups}
\title{Split data.frame/tibble based on grouping}
\usage{
split_groups(dframe)
}
\arguments{
\item{dframe}{data.frame}
}
\value{
list of data frames
}
\description{
Split data.frame/tibble based on grouping
}
