% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize}
\alias{standardize}
\title{Standardize numeric vector}
\usage{
standardize(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Standardized numeric vector (mean 0, sd 1)
}
\description{
Standardize numeric vector
}
\examples{
standardize(c(1,2,3,4,5))
}
