% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_columns.R
\name{standardize_columns}
\alias{standardize_columns}
\title{Standardize Columns}
\usage{
standardize_columns(data)
}
\arguments{
\item{data}{A data frame.}
}
\value{
A data frame with numeric columns standardized.
}
\description{
This function standardizes numeric columns in a dataset.
}
\examples{
data <- data.frame(a = c(1, 2, 3, 4), b = c(2, 3, 4, 5))
standardize_columns(data)
}
