% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{to_snake_case}
\alias{to_snake_case}
\title{Convert strings to snake_case}
\usage{
to_snake_case(data)
}
\arguments{
\item{data}{character Vector of strings to convert.}
}
\value{
character Vector in snake_case.
}
\description{
Convert strings to snake_case
}
