% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value_swap.R
\name{value_swap.data.frame}
\alias{value_swap.data.frame}
\title{Swaps values on two variables on a dataframe.}
\usage{
\method{value_swap}{data.frame}(x, .svp, .swp, .idx = NULL, ...)
}
\arguments{
\item{x}{A dataframe.}

\item{.svp}{The name of the first variable involved in the value swap.}

\item{.swp}{The name of the second variable involved in the value swap.}

\item{.idx}{An optional row index. If ommitted all values are swapped between variables.}

\item{...}{Arguments passed on to other methods.}
}
\description{
Swaps values of two variables on a dataframe. Takes an optional index.
}
\examples{

df <- data.frame(a=c(1, 2),
                 b=c(3, 4),
                 c=c(4, 6))

index <- c(TRUE, FALSE)

value_swap(df, "a", "b", index)

}
