% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value_swap.R
\name{value_swap.default}
\alias{value_swap.default}
\title{Swaps values on two inputs of equal length}
\usage{
\method{value_swap}{default}(x, ...)
}
\arguments{
\item{x}{Argument of a type for which no value_swap method exists yet.}

\item{...}{Arguments passed to other methods}
}
\description{
Fallback method
}
