% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Monnaie_Montants.R
\name{convert_currency_to_usd}
\alias{convert_currency_to_usd}
\title{Convertir les salaires en USD}
\usage{
convert_currency_to_usd(data)
}
\arguments{
\item{data}{Data frame contenant les colonnes salary, salary_currency, salary_in_usd et work_year}
}
\value{
Data frame avec les salaires convertis en USD
}
\description{
Convertit les salaires de différentes devises en USD en utilisant des taux de change historiques annuels
}
