% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacleanr_server.R
\name{datacleanr_server}
\alias{datacleanr_server}
\title{datacleanr server function}
\usage{
datacleanr_server(input, output, session, dataset, df_name, is_on_disk)
}
\arguments{
\item{input, output, session}{standard \code{shiny} boilerplate}

\item{dataset}{data.frame, tibble or data.table that needs cleaning}

\item{df_name}{character, name of dataset or file_path passed into shiny app}

\item{is_on_disk}{logical, whether df was read from file}
}
\description{
datacleanr server function
}
