% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{deduplicate_rows}
\alias{deduplicate_rows}
\title{Deduplicate data}
\usage{
deduplicate_rows(data, keys = NULL, keep = c("first", "last"))
}
\arguments{
\item{data}{data to deduplicate.}

\item{keys}{character|null Columns used as unique key; default NULL = all columns.}

\item{keep}{character Keep "first" or "last" occurrence when duplicates exist.}
}
\value{
deduplicated data.
}
\description{
Deduplicate data
}
