% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{enforce_types}
\alias{enforce_types}
\title{Enforce simple column types}
\usage{
enforce_types(data, num_threshold = 0.9, max_factor_levels = 20)
}
\arguments{
\item{data}{data.frame Input data}

\item{num_threshold}{numeric Proportion threshold for numeric conversion (default 0.9)}

\item{max_factor_levels}{integer Max distinct levels to convert to factor (default 20)}
}
\value{
data.frame Data with basic types enforced
}
\description{
Enforce simple column types
}
