% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extend_palette}
\alias{extend_palette}
\title{extend brewer palette}
\usage{
extend_palette(n)
}
\arguments{
\item{n}{numeric, number of colors}
}
\value{
color vector of length n
}
\description{
extend brewer palette
}
