% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_scoped}
\alias{filter_scoped}
\title{Apply filter based on a statement, scoped to groups}
\usage{
filter_scoped(dframe, statement, scope_at)
}
\arguments{
\item{dframe}{data.frame/tbl, grouped or ungrouped}

\item{statement}{character, statement for filtering (valid AND invalid)}

\item{scope_at}{numeric, group indices to apply filter statements to}
}
\value{
List, containing logical item stating \code{success}, and if \code{TRUE}
additional items \code{filtered_df} and the respective \code{expression_string} (character) for generating
the data.frame \ tibble
}
\description{
Apply filter based on a statement, scoped to groups
}
