% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso8601_char.R
\name{iso8601_char}
\alias{iso8601_char}
\title{Convert datetime string to ISO 8601 standard}
\usage{
iso8601_char(x, orders = NULL, tz = NULL)
}
\arguments{
\item{x}{(character) A vector of dates and times.}

\item{orders}{(character) Format of dates and times listed in x. Valid options are:
\itemize{
    \item{ymd}
    \item{ymd_H}
    \item{ymd_HM}
    \item{ymd_HMS}
    \item{ymd}
    \item{mdy_H}
    \item{mdy_HM}
    \item{mdy_HMS}
    \item{dmy}
    \item{dmy_H}
    \item{dmy_HM}
    \item{dmy_HMS}
}
Where:
\itemize{
    \item{y} is year
    \item{m} is month
    \item{d} is day
    \item{H} is hour in 24 hour clock
    \item{M} is minute
    \item{s} is second
}}

\item{tz}{(character) Time zone offset with respect to UTC (e.g. '+5', '-11')}
}
\value{
A vector of dates and times in the ISO 8601 standard to the precision of
    the input datetime value.
}
\description{
Convert a vector of dates and times to the standard ISO 8601 format
    YYYY-MM-DDThh:mm:ss. NOTE: Time zones are not yet supported.
}
