% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso8601_format.R
\name{iso8601_format}
\alias{iso8601_format}
\title{Get datetime format string from ISO 8601 datetimes}
\usage{
iso8601_format(x)
}
\arguments{
\item{x}{(character) A vector of dates and times created with
`EDIutils::iso8601_char`.}
}
\value{
(character) A datetime format string if only one is found, otherwise
    a vector of datetime format strings are returned.
}
\description{
Get format of ISO 8601 formatted character string.
}
\examples{
# Get format string
datetime <- iso8601_char(x = '2012-05-01 13:29:54', orders = 'ymd_HMS')
iso8601_format(datetime)
datetime <- iso8601_char(x = '2012-05-01 13:29:54', orders = 'ymd_HMS', tz = '-3')
iso8601_format(datetime)
datetime <- iso8601_char(x = '2012-05-01 13:29:54', orders = 'ymd_HMS', tz = '+5')
iso8601_format(datetime)

}
