% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso8601_get_format_string.R
\name{iso8601_get_format_string}
\alias{iso8601_get_format_string}
\title{Get the format specifier of ISO 8601 strings}
\usage{
iso8601_get_format_string(x)
}
\arguments{
\item{x}{(character) A vector ISO 8601 dates and times created with 
`dataCleanr::iso8601_convert`.}
}
\value{
(character) A date time format string representing user supplied ISO 
    8601 data. If more than one date time format is present, then the mode
    is returned.
}
\description{
Get the format string of user supplied ISO 8601 date times (e.g. 
    'YYYY-MM-DDThh:mm:ss).
}
\examples{
# Get format strings
iso8601_get_format_string('2012-05-01')
iso8601_get_format_string('2012-05-01T13')
iso8601_get_format_string('2012-05-01T13:29:54+05')

}
