% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso8601_read.R
\name{iso8601_read}
\alias{iso8601_read}
\title{Read ISO 8601 standard character string to datetime object (POSIXct, POSIXt)}
\usage{
iso8601_read(x)
}
\arguments{
\item{x}{(character) A vector of dates and times in the ISO8601 format:
\itemize{
    \item{YYYY-MM-DD}
    \item{YYYY-MM-DDThh}
    \item{YYYY-MM-DDThh:mm}
    \item{YYYY-MM-DDThh:mm:ss}
}}
}
\value{
A vector of datetime object (POSIXct, POSIXt).
}
\description{
Convert a character string in the ISO 8601 format to a datetime object
    (POSIXct, POSIXt). NOTE: Time zones are not currently supported.
}
