% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso8601_read.R
\name{iso8601_read}
\alias{iso8601_read}
\title{Read ISO 8601 standard character string to datetime object (POSIXct, POSIXt)}
\usage{
iso8601_read(x)
}
\arguments{
\item{x}{(character) A vector of dates and times created with 
`dataCleanr::iso8601_convert`.}
}
\value{
A vector of datetime object (POSIXct, POSIXt). If format is YYYY, then 
    integers are returned.
}
\description{
Convert a character string in the ISO 8601 format to a datetime object
    (POSIXct, POSIXt). NOTE: Time zones are not currently supported.
}
\examples{
# Read data strings created with iso8601_convert
datetimes <- iso8601_convert(x = c('2012-05-01 13:29:54', '2012-05-01 13:29', '1/5/2012 13'), orders = c('ymd_HMS', 'ymd_HM', 'dmy_H'))
datetimes <- iso8601_read(datetimes)
class(datetimes)
 

}
