% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso8601_read.R
\name{iso8601_read}
\alias{iso8601_read}
\title{Read ISO 8601 formatted strings into POSIXct, POSIXt}
\usage{
iso8601_read(x)
}
\arguments{
\item{x}{(character) A vector of ISO 8601 date times created with 
`dataCleanr::iso8601_convert`.}
}
\value{
(POSIXct, POSIXt) representation of date times, unless the format
    is 'YYYY', in which case integer class data is returned.
}
\description{
Read ISO 8601 formatted dates and times into POSIXct, POSIXt.
}
\details{
`iso8601_read` provides a lightweight option for reading ISO 8601 data 
    in the POSIXct POXIXt class. This function uses regular expressions to 
    extract the orders and time zone offset arguments and then passess the 
    info to `lubridate::parse_date_time` for parsing. Use 
    `lubridate::parse_date_time` for more options on handling time zone 
    offsets and locales.
}
\examples{
# Read date time strings
datetimes <- iso8601_read(c('2012-05-01T13:23:00', '2012-05-02T13:23:00'))
class(datetimes)

# Read date time strings with specified time zone offset
datetimes <- iso8601_read(c('2012-05-01T13:23:00-03', '2012-05-02T13:23:00-03'))
class(datetimes)
 

}
