% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_filter_str_select_box.R
\name{module_server_box_str_filter}
\alias{module_server_box_str_filter}
\title{Server Module: box for str filter condition}
\usage{
module_server_box_str_filter(input, output, session, selector, actionbtn)
}
\arguments{
\item{input, output, session}{standard}

\item{selector}{character, html selector for placement}

\item{actionbtn}{reactive, action button counter}
}
\description{
Server Module: box for str filter condition
}
