% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_filter_df.R
\name{module_server_df_filter}
\alias{module_server_df_filter}
\title{Server Module: filter info text and filtered df output}
\usage{
module_server_df_filter(input, output, session, dframe, condition_df)
}
\arguments{
\item{input, output, session}{standard \code{shiny} boilerplate}

\item{dframe}{data frame/tibble for filtering}

\item{condition_df}{data frame/tibble with filtering conditions and grouping scope}
}
\value{
df, either filtered or original, based on validity of \code{statements} in \code{condition_df}
}
\description{
Server Module: filter info text and filtered df output
}
