% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_extract_code_box.R
\name{module_server_extract_code}
\alias{module_server_extract_code}
\title{Server Module:  Selection Annotator}
\usage{
module_server_extract_code(
  input,
  output,
  session,
  df_label,
  filter_df,
  gvar,
  statements,
  sel_points,
  overwrite,
  is_on_disk,
  out_path
)
}
\arguments{
\item{input, output, session}{standard \code{shiny} boilerplate}

\item{df_label}{string, name of original df input}

\item{filter_df}{reactiveValue data frame with filter statements and scoping lvl}

\item{gvar}{reactive character, grouping vars for \code{dplyr::group_by}}

\item{statements}{reactive, lgl, vector of working statements}

\item{sel_points}{reactiveValue, data frame with selected point keys, annotations, and selection count}

\item{overwrite}{reacive value, TRUE/FALSE from checkbox input}

\item{is_on_disk}{Logical, whether df represented by \code{df_label} was on disk or from interactive \code{R} use}

\item{out_path}{reactive, List, with character strings providing directory paths and file names for saving/reading in code output}
}
\description{
Server Module:  Selection Annotator
}
