% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_str_select_box.R
\name{module_server_filter_str}
\alias{module_server_filter_str}
\title{Server Module: box for str filter condition}
\usage{
module_server_filter_str(input, output, session)
}
\arguments{
\item{input, output, session}{standard \code{shiny} boilerplate}
}
\description{
Server Module: box for str filter condition
}
\details{
provides UI text box element
}
