% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_group_selection.R
\name{module_server_group_select}
\alias{module_server_group_select}
\title{Server Module: group selection}
\usage{
module_server_group_select(input, output, session)
}
\arguments{
\item{input, output, session}{standard}
}
\description{
Server Module: group selection
}
