% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_group_selector_table.R
\name{module_server_group_selector_table}
\alias{module_server_group_selector_table}
\title{Server Module: box for str filter condition}
\usage{
module_server_group_selector_table(input, output, session, df)
}
\arguments{
\item{input, output, session}{standard \code{shiny} boilerplate}

\item{df}{data frame (either from overview or filtering tab)}
}
\description{
Server Module: box for str filter condition
}
\details{
provides UI text box element
}
