% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_data_summary.R
\name{module_server_summary}
\alias{module_server_summary}
\title{Server Module: data summary}
\usage{
module_server_summary(
  input,
  output,
  session,
  dframe,
  df_label,
  start_clicked,
  group_var_check
)
}
\arguments{
\item{input, output, session}{standard \code{shiny} boilerplate}

\item{dframe}{reactive, input data frame}

\item{df_label}{character, name of initial data set}

\item{start_clicked}{reactive holding start action button}

\item{group_var_check}{reactive holding group check output}
}
\description{
Server Module: data summary
}
