% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module_str_select_box.R
\name{module_ui_box_str_filter}
\alias{module_ui_box_str_filter}
\title{UI Module: box for str filter condition}
\usage{
module_ui_box_str_filter(id, actionbtn)
}
\arguments{
\item{id}{Character, identifier for variable selection}

\item{df}{data.frame loaded into gadget; should support df, tibble, data.table}
}
\description{
UI Module: box for str filter condition
}
