% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalisation.R
\name{normalize_all}
\alias{normalize_all}
\title{Normalize everything of an salary_tbl}
\usage{
normalize_all(data)
}
\arguments{
\item{data}{the salary_tbl to normalize}
}
\value{
data_normalized the normalized salary_tbl
}
\description{
Normalize everything of an salary_tbl
}
