% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalisation.R
\name{normalize_factor}
\alias{normalize_factor}
\title{Normalize a vector to a factor
(use with \code{df$col <- normalize_factor(df$col, mapping, levels)})}
\usage{
normalize_factor(vector, mapping, levels_in_factor, ordered_factor = FALSE)
}
\arguments{
\item{vector}{The raw vector to normalise.}

\item{mapping}{A named vector (key-value) indicating the transformation
(e.g. \code{c("old_val" = "new_val")}).}

\item{levels_in_factor}{A vector containing all valid levels for the final factor.}

\item{ordered_factor}{Logical (default \code{FALSE}). If \code{TRUE}, creates an ordered factor.}
}
\value{
A normalised \code{factor} vector.
}
\description{
Transform a raw vector into a factor by applying a mapping table.
Values that are not found or are missing are replaced by \code{"Unknown"}.
}
\details{
This function is useful to standardise poorly formatted categorical columns.
It performs the following operations:
\enumerate{
\item Replace the values of the vector according to the supplied \code{mapping}.
\item Replace \code{NA} values with the string \code{"Unknown"}.
\item Convert the result to a \code{factor} with the imposed levels (\code{levels_in_factor}).
}
}
\examples{
raw_data <- c("FR", "US", "UK", "FR", "XX")

# Mapping table
map_countries <- c(
  "FR" = "France",
  "US" = "United States",
  "UK" = "United Kingdom"
)

# Allowed levels
valid_levels <- c("France", "United States", "United Kingdom", "Unknown")

# Apply normalisation
normalize_factor(raw_data, map_countries, valid_levels)
}
\seealso{
Other Normalization functions: 
\code{\link{normalize_all}()},
\code{\link{normalize_remote_ratio}()}
}
\concept{Normalization functions}
