% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalisation.R
\name{normalize_factor}
\alias{normalize_factor}
\title{Normalize a vector to a factor. (use with "df$col <- normalize_to_factor(df$col, mapping, levels)")}
\usage{
normalize_factor(vector, mapping, levels_in_factor, ordered_factor = FALSE)
}
\arguments{
\item{vector}{the vector to normalize}

\item{mapping}{the mapping to indicate which values to transform in which.}

\item{levels_in_factor}{les niveaux à indiquer}

\item{ordered_factor}{Indicate if the vector has to be ordonned, FALSE by default}
}
\value{
normalized the normalized vector
}
\description{
Normalize a vector to a factor. (use with "df$col <- normalize_to_factor(df$col, mapping, levels)")
}
