% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalisation.R
\name{normalize_remote_ratio}
\alias{normalize_remote_ratio}
\title{titre}
\usage{
normalize_remote_ratio(data, binary = FALSE, threshold = 50)
}
\arguments{
\item{data}{data.frame contenant la colonne \code{remote_ratio}}

\item{binary}{logique, si TRUE convertit en 0/100 (par défaut FALSE)}

\item{threshold}{seuil pour le binaire (défaut 50)}
}
\value{
data.frame avec \code{remote_ratio} corrigée
}
\description{
titre
}
