% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{read_raw_csv}
\alias{read_raw_csv}
\title{Import CSV as data.frame}
\usage{
read_raw_csv(file_path)
}
\arguments{
\item{file_path}{chemin vers le CSV}
}
\value{
data.frame
}
\description{
Import CSV as data.frame
}
