% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_outliers.R
\name{remove_outliers}
\alias{remove_outliers}
\title{Remove outliers with IQR}
\usage{
remove_outliers(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Numeric vector with outliers set to NA
}
\description{
Remove outliers with IQR
}
\examples{
remove_outliers(c(1,2,3,100))
}
