% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_groups}
\alias{split_groups}
\title{Identify filter inputs by id and label}
\usage{
split_groups(dframe)
}
\arguments{
\item{allinputs}{list of all currently active inputs}
}
\value{
list with id of last filter (character), and number of last filter (numeric)
}
\description{
Identify filter inputs by id and label
}
