% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{standardize_colnames}
\alias{standardize_colnames}
\title{Convert strings to snake_case}
\usage{
standardize_colnames(data)
}
\arguments{
\item{data}{character Vector of strings to convert.}
}
\value{
character Vector in snake_case.
}
\description{
Convert strings to snake_case
}
