% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_schema}
\alias{validate_schema}
\title{Test the presence of required colums in a dataframe}
\usage{
validate_schema(dataframe, required_colums)
}
\arguments{
\item{dataframe}{Dataframe to test}

\item{required_colums}{vector with the required colums names to test in dataframe}
}
\value{
presence Indication to the user on the preence of the colums
}
\description{
Test the presence of required colums in a dataframe
}
